feat: implemented add and retrieve pocket

+++ src/main/java/com/montrack/montrack/wallet/model/dto/WalletDto.java
@@ -11,6 +11,6 @@ public class WalletDto {
     private Long id;
     private String name;
     private String currency;
-    private Long balance;
+    private int balance;
     private boolean isActive;
 }
